Initial commit
Very early experiment towards a Wikipedia/Wikidata geolocation
and navigation helper tool for Microsoft Flight Simulator 2020.

Currently connects to the simulator and queries every second for
the current altitude, latitude, and longitude of the user's plane.

In future, will provide a nice user interface for using that data
to query for Wikipedia articles and interesting items in Wikidata
that can be displayed on a map in the app and/or marked within the
simulator.

This is not an official project of the Wikimedia Foundation, and is
just a fun experiment for learning about the MSFS SDK and SimConnect,
so don't get too excited! But have fun and learn something cool. :D